Buffers in pool may not always point at the Metal video compositors output
We received a report of black frames in an applications video content once it was encoded with NextLevel.

Investigation revealed a similar issue as https://github.com/NextLevel/NextLevelSessionExporter/issues/38.

It appears NextLevelSessionExporter is pulling buffers from the pool that was created, and sending those buffers back to the API. The  assumption a buffer pulled from that pool will contain the output of the Metal video compositor is not always correct.

This change will pass the pixelBuffer directly, or use the toBuffer if a renderHandler exists.